requirements.txt
xkits-logger>=0.2
xkits-lib>=0.3
pyOpenSSL
